feat : Main 컴포넌트에서 CommentItem 분리 및 댓글 삭제 추가
