+++ src/pages/jeongdo/Main/CommentItem/Comment.scss
@@ -1,0 +1,26 @@
+.post__comment__guest-Wrap {
+  font-size: 12px;
+}
+
+.post__myComment {
+  font-size: 12px;
+}
+
+.post__myComment a {
+  margin-right: 6px;
+}
+
+.rmBack {
+  background-color: rgba(255, 255, 255, 0);
+  color: #a4a4a4;
+  border: none;
+  outline: none;
+  cursor: pointer;
+}
+.on {
+  color: #ed4956;
+}
+
+.fa-heart {
+  cursor: pointer;
+}
