+++ src/pages/jeongdo/Main/Main.scss
@@ -133,26 +133,6 @@ a {
   margin-right: 5px;
 }
 
-.post__comment__guest-Wrap {
-  font-size: 12px;
-}
-
-.post__myComment {
-  font-size: 12px;
-}
-
-.post__myComment a {
-  margin-right: 6px;
-}
-
-.rmBack {
-  background-color: rgba(255, 255, 255, 0);
-  color: #a4a4a4;
-  border: none;
-  outline: none;
-  cursor: pointer;
-}
-
 span a {
   font-weight: bold;
 }
